pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = ""
authors = [
    {name = "a.kuzmenko",email = "a.kuzmeko@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
